backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
celery==5.3.6
redis==5.0.1
psycopg2-binary==2.9.9
minio==7.2.5
python-multipart==0.0.9
pydantic==2.6.3
pydantic-settings==2.2.1
SQLAlchemy==2.0.25
openai==1.14.3
python-slugify==8.0.4
